requirements.txt
customtkinter==5.1.2
matplotlib==3.7.1
moviepy==1.0.3
numpy==1.24.2
Pillow==9.5.0
Pillow==10.0.1
tkVideoPlayer==2.3
